refactor(storage): make msg independent from state
